合并

+++ src/main/java/cn/edu/nju/movietubeserver/dao/RateDetailDao.java
@@ -18,6 +18,6 @@ public interface RateDetailDao
 
     int updateRateDetailById(@Param("rateDetailPo") RateDetailPo rateDetailPo);
 
-    Double getMyRateByMovieId(@Param("userId") Integer userId, @Param("movieId") Long movieId);
+    Integer getMyRateByMovieId(@Param("userId") Integer userId, @Param("movieId") Long movieId);
 
 }
